refactor(connectors): D4 — split lazy-mesh.ts into namespace directory
Replaces the 1428-LOC packages/gateway/src/connectors/lazy-mesh.ts with
a directory of concern-focused sibling files:

- drain.ts                    — LazyDrainTracker class (33 LOC)
- index.ts                    — re-export shim (3 LOC)
- keys.ts                     — LAZY_MESH constants + path helpers
                                (38 LOC; MCP_CONNECTORS_ROOT path
                                 adjusted: extra ".." to compensate
                                 for new directory depth — only
                                 non-verbatim line in the move)
- slot.ts                     — LazyMcpSlot + ServerSpec +
                                MeshSpawnContext interface (37 LOC)
- tool-map.ts                 — LazyMeshToolMap, MeshLogger, helpers
                                (45 LOC)
- user-mcp.ts                 — recordArgsJsonFailure +
                                ensureUserMcpClient (80 LOC)
- credential-orchestration.ts — 11 file-private wrappers +
                                ensureCredentialConnectorsRunning
                                (152 LOC)
- phase3-config.ts            — 8 phase3AddXxxMcp helpers +
                                buildPhase3Servers (181 LOC)
- mesh.ts                     — LazyConnectorMesh class with
                                constructor-bound spawnContext +
                                16 thin shells + tool-aggregation +
                                factory (432 LOC)
- connector-spawns.ts         — 16 ensureXxxMcp free functions
                                taking MeshSpawnContext (627 LOC)

Mostly-mechanical move; zero behavioral change. Per-connector spawn
methods become free functions accepting a constructor-bound
MeshSpawnContext interface, letting them live in sibling files
without exposing the slot state-machine on the public class API.
The LazyConnectorMesh class stays the state owner; its 16
ensureXxxRunning methods become 1-line shells that delegate to the
free functions. ensureUserMcpConnectorsRunning stays on the class
because it iterates this.lazySlots.keys() directly.

Largest resulting file is connector-spawns.ts at 627 LOC, well under
the 800 D4 threshold.

Ten consumer files updated their import path from "./lazy-mesh.ts"
(or "../connectors/lazy-mesh.ts" / "../../connectors/lazy-mesh.ts")
to the corresponding ".../lazy-mesh/index.ts".

Updated invariant wiring per CLAUDE.md "update test + docs in same
commit" rule:
- packages/gateway/src/security-invariants.test.ts — I1 tests now
  scan the lazy-mesh/ directory (concatenated source) instead of
  reading lazy-mesh.ts.
- packages/gateway/src/connectors/lazy-mesh-args-json.test.ts — S8-F8
  Date.now()-absence test now scans the directory.
- CLAUDE.md, GEMINI.md, docs/SECURITY-INVARIANTS.md,
  docs/architecture.md — I1 wired-at + Key File Locations rows
  reference the new directory.

Spec: docs/superpowers/specs/2026-05-02-d4-lazy-mesh-split-design.md
Plan: docs/superpowers/plans/2026-05-02-d4-lazy-mesh-split.md
