+++ packages/gateway/src/connectors/lazy-mesh/drain.ts
@@ -1,0 +1,33 @@
+/**
+ * S8-F7 — per-slot in-flight refcount with awaitable drain.
+ * Used by LazyConnectorMesh to defer disconnect while tool calls are running.
+ */
+export class LazyDrainTracker {
+  private inFlight = 0;
+  private resolveDrained: (() => void) | undefined;
+  private drained: Promise<void> | undefined;
+
+  bump(): void {
+    this.inFlight += 1;
+    this.drained ??= new Promise<void>((r) => {
+      this.resolveDrained = r;
+    });
+  }
+
+  drop(): void {
+    if (this.inFlight > 0) this.inFlight -= 1;
+    if (this.inFlight === 0 && this.resolveDrained !== undefined) {
+      this.resolveDrained();
+      this.drained = undefined;
+      this.resolveDrained = undefined;
+    }
+  }
+
+  awaitDrain(): Promise<void> {
+    return this.drained ?? Promise.resolve();
+  }
+
+  get count(): number {
+    return this.inFlight;
+  }
+}
