+++ packages/gateway/src/connectors/lazy-mesh/index.ts
@@ -1,0 +1,3 @@
+export { LazyDrainTracker } from "./drain.ts";
+export { createLazyConnectorMesh, LazyConnectorMesh } from "./mesh.ts";
+export { type MeshLogger, mergeToolMapsOrThrow } from "./tool-map.ts";
